working save all transactions to db
